Add first draft of on-screen keyboard
The on-screen keyboard can be enabled or disabled in the main menu. When you press a key, the keypress is simulated using a java.awt.Robot. Keys that need to be held like Shift or Control need two clicks: One for pressing and one for releasing. The keyboard layout is loaded from a text file.