Added validation for the shortcut name
The shortcut name can only contain letters, numbers and hyphens, has to be smaller than 20 characters and isn't allowed to start with a hyphen. You now have to right click a shortcut to start it. Added a delay to the clicks. They are executed at the same speed and the same delay as they got recorded with.